requirements.txt
# === Core AI / LLM ===
anthropic>=0.25.0
langchain>=0.2.0
langchain-anthropic>=0.1.0
langgraph>=0.1.0

# === Translation ===
# IndicTrans2 via IndicTransToolkit (quantized, lighter than raw model)
# Install separately if using local model:
#   pip install git+https://github.com/AI4Bharat/IndicTransToolkit.git
# For MVP, Claude API handles translation (USE_CLAUDE_FOR_TRANSLATION=true)
langdetect>=1.0.9

# === Backend ===
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9    # For image file uploads
httpx>=0.27.0              # Async HTTP client
pydantic>=2.0.0
pydantic-settings>=2.0.0  # For .env loading

# === Twilio (WhatsApp + SMS) ===
twilio>=9.0.0

# === Frontend ===
streamlit>=1.35.0
Pillow>=10.3.0             # Image handling

# === Testing ===
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0

# === Dev / Code Quality ===
black>=24.0.0
python-dotenv>=1.0.0
